pyproject.toml
[tool.poetry]
name = "retail-network"
version = "0.1.0"
description = ""
authors = ["Josef Vas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
setuptools = "^73.0.1"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
